+++ src/Benchmarks/ParallelHeapBenchmarks.cs
@@ -12,7 +12,7 @@ namespace Benchmarks
         private ThreadParallelRunner<ClrSegment> _runner;
 
         [ParamsSource(nameof(CacheSizeSource))]
-        public int CacheSize { get; set; }
+        public long CacheSize { get; set; }
 
         [ParamsSource(nameof(OSMemoryFeaturesSource))]
         public bool UseOSMemoryFeatures { get; set; }
@@ -21,9 +21,8 @@ namespace Benchmarks
         public int Threads { get; set; }
 
         public IEnumerable<bool> OSMemoryFeaturesSource => BenchmarkSwitches.OSMemoryFeatureFlags;
-        public IEnumerable<ulong> CacheSizeSource => BenchmarkSwitches.RelevantCacheSizes;
+        public IEnumerable<long> CacheSizeSource => BenchmarkSwitches.RelevantCacheSizes;
         public IEnumerable<int> ThreadCountSource => BenchmarkSwitches.Parallelism;
-
 
         [GlobalSetup]
         public void Setup()
